+++ tests/e2e/builder/cases/top-level-await/src/index.ts
@@ -1,0 +1,11 @@
+import { sayHello } from './sayHello';
+
+const result = await sayHello();
+
+declare global {
+  interface Window {
+    foo: string;
+  }
+}
+
+window.foo = result;