+++ tests/scripts/dev-project.sh
@@ -1,2 +1,0 @@
-#!/usr/bin/env bash
-pnpm --parallel --filter "@cypress-test/*" run dev