+++ packages/builder/builder-webpack-provider/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @modern-js/builder-webpack-provider
 
+## 2.25.1
+
+### Patch Changes
+
+- Updated dependencies [b10045f]
+- Updated dependencies [9f78d0c]
+  - @modern-js/builder-shared@2.25.1
+  - @modern-js/utils@2.25.1
+  - @modern-js/babel-preset-app@2.25.1
+  - @modern-js/babel-preset-base@2.25.1
+  - @modern-js/server@2.25.1
+  - @modern-js/types@2.25.1
+
 ## 2.25.0
 
 ### Patch Changes