+++ packages/builder/plugin-image-compress/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @modern-js/builder-plugin-image-compress
 
+## 2.25.1
+
+### Patch Changes
+
+- Updated dependencies [9f78d0c]
+  - @modern-js/utils@2.25.1
+
 ## 2.25.0
 
 ### Patch Changes