+++ packages/builder/plugin-stylus/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @modern-js/builder-plugin-stylus
 
+## 2.25.1
+
+### Patch Changes
+
+- Updated dependencies [b10045f]
+  - @modern-js/builder-shared@2.25.1
+  - @modern-js/builder-rspack-provider@2.25.1
+  - @modern-js/builder-webpack-provider@2.25.1
+
 ## 2.25.0
 
 ### Patch Changes