+++ packages/cli/plugin-swc/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @modern-js/core
 
+## 2.25.1
+
+### Patch Changes
+
+- Updated dependencies [9f78d0c]
+  - @modern-js/utils@2.25.1
+  - @modern-js/builder-plugin-swc@2.25.1
+
 ## 2.25.0
 
 ### Patch Changes