+++ packages/generator/generator-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js/generator-common
 
+## 3.1.26
+
+### Patch Changes
+
+- @modern-js/plugin-i18n@2.25.1
+
 ## 3.1.25
 
 ### Patch Changes