+++ packages/generator/generators/server-generator/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @modern-js/server-generator
 
+## 3.1.26
+
 ## 3.1.25
 
 ## 3.1.24