+++ packages/server/plugin-server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @modern-js/plugin-server
 
+## 2.25.1
+
+### Patch Changes
+
+- Updated dependencies [9f78d0c]
+  - @modern-js/utils@2.25.1
+  - @modern-js/server-utils@2.25.1
+
 ## 2.25.0
 
 ### Patch Changes