+++ packages/server/utils/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @modern-js/server-utils
 
+## 2.25.1
+
+### Patch Changes
+
+- Updated dependencies [9f78d0c]
+  - @modern-js/utils@2.25.1
+  - @modern-js/babel-preset-base@2.25.1
+  - @modern-js/babel-compiler@2.25.1
+
 ## 2.25.0
 
 ### Patch Changes