+++ packages/toolkit/remark-container/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @modern-js/remark-container
 
+## 2.25.1
+
 ## 2.25.0
 
 ## 2.24.0