+++ packages/toolkit/types/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @modern-js/types
 
+## 2.25.1
+
 ## 2.25.0
 
 ### Minor Changes