+++ scripts/lint-package-json/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @scripts/lint-package-json
 
+## 2.25.1
+
 ## 2.25.0
 
 ## 2.24.0